+++ src/lib/theme/mobileapp/index.ts
@@ -1,9 +1,9 @@
 import mainHTML from './html/main.html?raw';
 import { generateAttributes, generateTabButtons, useI18n } from '../../../utils'
 import { template } from "../../../utils/template";
-import { PaymentProvider, type FormConfig, type FormSetting, type FooterInfo } from '../../types';
-import { createCreditCardForm, createFooter, createPaymentHeader, prepareDiscountSection } from '../../create'
-import { getPackageName, getQuantityInfo } from '../../../utils/getPackageInfo';
+import { PaymentProvider, type FormConfig, type FormSetting, type FooterInfo, PackageType, PackageCondition } from '../../types';
+import { createCreditCardForm, createFooter, createPaymentHeader, createPriceTable, prepareDiscountSection } from '../../create'
+import { getQuantityInfo } from '../../../utils/getPackageInfo';
 import { prepareProvider } from './utils';
 
 export function generateThemeMobileApp(params: {
@@ -47,7 +47,7 @@ export function generateThemeMobileApp(params: {
 
   const totalPrice = config.packageInfo?.totalPayableAmount || '0.00 USD';
   const packagePrice = config.packageInfo?.discount.original;
-  const additionalPrice = config.packageInfo?.discount.price;
+  const additionalPrice = `0.00 ${config.general.currency}`;
 
   if (providerButtons) {
     primaryProvider += `<div class="zotlo-checkout__seperator"><span>${$t('common.orAnotherWay')}</span></div>`
@@ -55,10 +55,8 @@ export function generateThemeMobileApp(params: {
 
   const hasProductConfig = Object.prototype.hasOwnProperty.call(config.design, 'product');
   const showProductImage = hasProductConfig && Object.prototype.hasOwnProperty.call(config.design.product, 'productImage') ? !!config.design?.product?.productImage?.show : true;
-  const showSubtotal = hasProductConfig && Object.prototype.hasOwnProperty.call(config.design.product, 'showSubtotalText') ? !!config.design?.product?.showSubtotalText : true;
   const showAdditonalText = hasProductConfig && Object.prototype.hasOwnProperty.call(config.design.product, 'additionalText') ? !!config.design?.product?.additionalText?.show : true;
   const productImage = showProductImage ? (config.general.productImage || config.design?.product?.productImage.url || '') : '';
-  const packageName = getPackageName(config);
   const additionalText = showAdditonalText
     ? (
       config.general.additionalText ||
@@ -71,6 +69,13 @@ export function generateThemeMobileApp(params: {
 
   const paymentHeader = createPaymentHeader({ config });
   const footer = createFooter(footerInfo) || '';
+  let priceTable = '';
+  
+  if (config.packageInfo?.condition !== PackageCondition.ONETIME_PAYMENT) {
+    priceTable = createPriceTable({ config }) || '';
+  }
+
+  const isSubscription = config.paymentData?.package?.packageType === PackageType.SUBSCRIPTION;
 
   return template(mainHTML, {
     DIR: dir,
@@ -80,18 +85,18 @@ export function generateThemeMobileApp(params: {
       ...(config.cardUpdate ? {'data-type': 'card'} : {})
     }),
     HEADER: paymentHeader || '',
+    PRICE_TABLE: priceTable || '',
+    SHOW_TOTAL: !isSubscription,
     PACKAGE_SUMMARY: !config.cardUpdate,
     PACKAGE_IMAGE: productImage,
-    PACKAGE_NAME: packageName,
     PACKAGE_PRICE: packagePrice,
-    SHOW_SUBTOTAL: !!packageName && showSubtotal,
     STATIC_SUBTOTAL: $t('common.subtotal'),
     STATIC_TOTAL: $t('common.totalDue'),
     ADDITIONAL_TEXT: additionalText,
     ADDITIONAL_PRICE: additionalPrice,
     TOTAL_PRICE: totalPrice,
     QUANTITY_INFO: getQuantityInfo(config),
-    DISCOUNT_SECTION: prepareDiscountSection({ config }),
+    DISCOUNT_SECTION: isSubscription ? '' : prepareDiscountSection({ config }),
     PRIMARY_PROVIDER: primaryProvider,
     TAB_BUTTONS: tabButtons,
     PROVIDERS: providerButtons,