+++ src/utils/getPackageInfo.ts
@@ -1,4 +1,4 @@
-import { type FormConfig, PackageInfoType, PackageType, TrialPackageType } from "../lib/types";
+import { type FormConfig, PackageCondition, PackageInfoType, PackageType, TrialPackageType } from "../lib/types";
 import { useI18n, getIsDiscountCodeApplied } from "../utils";
 import { template } from "./template";
 
@@ -28,11 +28,16 @@ export function getPackageInfo(config?: FormConfig): PackageInfoType {
   };
 }
 
-export function getQuantityInfo(config: FormConfig) {
+export function getQuantityInfo(config: FormConfig, onlyUnits?: boolean) {
   const { $t } = useI18n(config.general.localization);
   const quantity = config?.settings?.quantitySetting?.quantity || 1;
   const templateParams = getPackageTemplateParams(config);
   if (+quantity <= 1) return "";
+
+  if (onlyUnits) {
+    return template($t('form.quantity.includesNumberUnits'), templateParams)
+  }
+
   return template($t('form.quantity.info'), templateParams);
 }
 
@@ -167,11 +172,11 @@ function getPackageCondition(config?: FormConfig, options?: { isTrialUsed?: bool
   const { isOneTimePayment, isNoTrial, isTrialUsed:isTrialUsedValue } = getPackageTypeConditions(config || {} as FormConfig, { isTrialUsed });
 
   if (isOneTimePayment) {
-    condition = 'onetime_payment';
+    condition = PackageCondition.ONETIME_PAYMENT;
   } else if (isNoTrial) {
-    condition = 'plan_with_no_trial'
+    condition = PackageCondition.PLAN_WITH_NO_TRIAL;
   } else if (isTrialUsedValue) {
-    condition = 'package_with_trial_used';
+    condition = PackageCondition.PACKAGE_WITH_TRIAL_USED;
   }
 
   return condition;
@@ -204,6 +209,33 @@ export function getPackageTemplateParams(config: FormConfig) {
   const { period = 0, trialPeriod = 0, periodType, trialPeriodType } = packageInfo || {};
   const { $t } = useI18n(config.general.localization);
 
+  let periodLabel = '';
+  let periodNaming = '';
+  let trialPeriodLabel = '';
+  let trialPeriodNaming = '';
+
+
+  if (period !== 0) {
+    periodLabel = $t(`common.periodBase.${periodType}`, { count: period });
+
+    if (period === 1) {
+      periodNaming = $t(`common.periodNaming.${periodType}`);
+    } else {
+      periodNaming = periodLabel;
+    }
+  }
+
+  if (trialPeriod !== 0) {
+    trialPeriodLabel = $t(`common.periodBase.${trialPeriodType}`, { count: trialPeriod })
+
+    if (trialPeriod === 1) {
+      periodNaming = $t(`common.periodNaming.${trialPeriodType}`);
+    } else {
+      trialPeriodNaming = trialPeriodLabel;
+    }
+  }
+
+
   return {
     QUANTITY: config?.settings?.quantitySetting?.quantity || 1,
     UNIT_PRICE: packageInfo?.totalPayableBaseAmount || "",
@@ -220,8 +252,12 @@ export function getPackageTemplateParams(config: FormConfig) {
     TRIAL_PRICE: packageInfo?.trialPrice || "",
     DAILY_PRICE: packageInfo?.dailyPrice || "",
     WEEKLY_PRICE: packageInfo?.weeklyPrice || "",
-    PERIOD: period === 0 ? '' : $t(`common.periods.${periodType}`, { count: period }),
-    TRIAL_PERIOD: period === 0 ? '' : $t(`common.periods.${trialPeriodType}`, { count: trialPeriod }),
+    PERIOD: periodLabel,
+    PERIOD_TYPE: periodType,
+    PERIOD_NAMING: periodNaming,
+    TRIAL_PERIOD: trialPeriodLabel,
+    TRIAL_PERIOD_TYPE: trialPeriodType,
+    TRIAL_PERIOD_NAMING: trialPeriodNaming
   };
 }